rendering to two windows works now, you can use multiple simulators now. next x button on windows doesn't work.

+++ include/patterns/patterns.h
@@ -27,7 +27,7 @@ public:
     void process();
 private:
     static Pattern_t *patterns;
-    static int num_patterns;
+    static size_t num_patterns;
 };
 
 #endif
